requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
bcrypt==3.2.0
cffi==1.17.1
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.11
fastcrud==0.15.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
passlib==1.7.4
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
python-dotenv==1.0.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
SQLAlchemy-Utils==0.41.2
starlette==0.46.0
typing_extensions==4.12.2
uvicorn==0.34.0
jwt==1.3.1
python-dotenv==1.0.1

alembic==1.15.1